+++ src/Exceptions/IpRestrictAccessException.php
@@ -2,6 +2,9 @@
 
 namespace Solital\Core\Exceptions;
 
+/**
+ * @deprecated Use trigger_error()
+ */
 class IpRestrictAccessException extends \Exception
 {
 }