+++ src/Http/Middleware/MiddlewareCallableInterface.php
@@ -2,6 +2,9 @@
 
 namespace Solital\Core\Http\Middleware;
 
+/**
+ * @deprecated Use Solital\Core\Http\Middleware\BaseMiddlewareInterface
+ */
 interface MiddlewareCallableInterface
 {