+++ src/Http/Middleware/MiddlewareInterface.php
@@ -4,6 +4,9 @@ namespace Solital\Core\Http\Middleware;
 
 use Solital\Core\Http\Request;
 
+/**
+ * @deprecated Use Solital\Core\Http\Middleware\BaseMiddlewareInterface
+ */
 interface MiddlewareInterface
 {
     /**