+++ src/Kernel/Traits/KernelTrait.php
@@ -21,7 +21,7 @@ trait KernelTrait
     /**
      * That variables must be changed manually
      */
-    const SOLITAL_VERSION   = "4.0.0-rc1";
+    const SOLITAL_VERSION   = "4.0.1";
     const SITE_DOC_DOMAIN   = "https://solital.github.io/site/";
     const DEBUG             = false;
     const DEBUG_DATABASE    = false;